Fix overflowing inputs on some platforms.
`width: 100%` sets the content-box width, unless the box-sizing value is
`border-box`, see:

  https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing

(`box-sizing: border-box` is already the default for `<select>`).

In platforms with more padding by default on inputs (like Firefox on Linux for
example) this can cause the inputs to overflow the popup. Also shrink the
padding manually to what the popup layout expects, so that it doesn't grow past
the popup.

Additionally we remove the appearance from checkboxes so that they don't take up
space.

This fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1625716.