note: explain why redirect is needed for redd.it
If you try visiting a redd.it site like <https://redd.it/1fwqjw> (first
link that came up when searching "site:redd.it"), you may think that the
redirect works because it goes to your configured instance.  However,
there's a leak: redd.it is a redirect to a www.reddit.com site, and
that's when the redirect to instance happens.  If you observe the
network traffic, you'll see that a GET request goes to redd.it, and
redd.it returns HTTP code 302 and location https://www.reddit.com/...

Let's not do that.  Redirect at redd.it so that requests don't hit
reddit servers.  This does not seem to work for teddit instances, so
don't redirect for them.